refactor: move openaicompat and anthropicwire under providers/internal
Both packages are used only by providers/*, so relocate them from internal/ to
providers/internal/ to co-locate with their consumers. They stay internal
(now scoped to the providers subtree). Import paths updated; no logic change.
