+++ providers/internal/anthropicwire/anthropicwire.go
