+++ providers/internal/anthropicwire/stream.go
