+++ providers/internal/anthropicwire/stream_test.go
