+++ providers/internal/openaicompat/complete.go
