+++ providers/internal/openaicompat/complete_test.go
