+++ providers/internal/openaicompat/embedding.go
