+++ providers/internal/openaicompat/embedding_test.go
