+++ providers/internal/openaicompat/request.go
