+++ providers/internal/openaicompat/request_test.go
