+++ providers/internal/openaicompat/stream.go
