+++ providers/internal/openaicompat/stream_test.go
